Module diem_rest_client::dpn

Structs

AccountBalance

Balance

BalanceResource

The balance resource held under an account.

Configuration

ConfigurationResource

Diem

DiemConfig

DiemVersion

OnChainDiemVersion

Defines the version of Diem Validator software.

Constants

CORE_CODE_ADDRESS

Traits

OnChainConfig

Trait to be implemented by a Rust struct representation of an on-chain config that is stored in storage as a serialized byte array

Functions

config_struct_tag

diem_root_address

diem_version_identifier